refactor eventList, move filer to new component
